+++ src/test/java/edu/harvard/iq/dataverse/DatasetVersionTest.java
@@ -86,117 +86,5 @@ public class DatasetVersionTest {
         assertFalse(nonDraft.isInReview());
     }
 
-    /**
-     * See also SchemaDotOrgExporterTest.java for more extensive tests.
-     */
-    @Test
-    public void testGetJsonLd() throws ParseException {
-        Dataset dataset = new Dataset();
-        License license = new License("CC0 1.0", "You can copy, modify, distribute and perform the work, even for commercial purposes, all without asking permission.", URI.create("http://creativecommons.org/publicdomain/zero/1.0"), URI.create("/resources/images/cc0.png"), true);
-        license.setDefault(true);
-        dataset.setProtocol("doi");
-        dataset.setAuthority("10.5072/FK2");
-        dataset.setIdentifier("LK0D1H");
-        DatasetVersion datasetVersion = new DatasetVersion();
-        datasetVersion.setDataset(dataset);
-        datasetVersion.setVersionState(DatasetVersion.VersionState.DRAFT);
-        assertEquals("", datasetVersion.getPublicationDateAsString());
-        // Only published datasets return any JSON.
-        assertEquals("", datasetVersion.getJsonLd());
-        datasetVersion.setVersionState(DatasetVersion.VersionState.RELEASED);
-        datasetVersion.setVersionNumber(1L);
-        SimpleDateFormat dateFmt = new SimpleDateFormat("yyyyMMdd");
-        Date publicationDate = dateFmt.parse("19551105");
-        datasetVersion.setReleaseTime(publicationDate);
-        dataset.setPublicationDate(new Timestamp(publicationDate.getTime()));
-        Dataverse dataverse = new Dataverse();
-        dataverse.setName("LibraScholar");
-        dataset.setOwner(dataverse);
-        TermsOfUseAndAccess terms = new TermsOfUseAndAccess();
-        terms.setLicense(license);
-        datasetVersion.setTermsOfUseAndAccess(terms);
-        String jsonLd = datasetVersion.getJsonLd();
-        logger.fine("jsonLd: " + JsonUtil.prettyPrint(jsonLd));
-        JsonReader jsonReader = Json.createReader(new StringReader(jsonLd));
-        JsonObject obj = jsonReader.readObject();
-        assertEquals("http://schema.org", obj.getString("@context"));
-        assertEquals("Dataset", obj.getString("@type"));
-        assertEquals("https://doi.org/10.5072/FK2/LK0D1H", obj.getString("@id"));
-        assertEquals("https://doi.org/10.5072/FK2/LK0D1H", obj.getString("identifier"));
-        assertEquals(null, obj.getString("schemaVersion", null));
-        assertEquals("http://creativecommons.org/publicdomain/zero/1.0", obj.getString("license"));
-        assertEquals("1955-11-05", obj.getString("dateModified"));
-        assertEquals("1955-11-05", obj.getString("datePublished"));
-        assertEquals("1", obj.getString("version"));
-        // TODO: if it ever becomes easier to mock a dataset title, test it.
-        assertEquals("", obj.getString("name"));
-        // TODO: If it ever becomes easier to mock authors, test them.
-        JsonArray emptyArray = Json.createArrayBuilder().build();
-        assertEquals(emptyArray, obj.getJsonArray("creator"));
-        assertEquals(emptyArray, obj.getJsonArray("author"));
-        // TODO: If it ever becomes easier to mock subjects, test them.
-        assertEquals(emptyArray, obj.getJsonArray("keywords"));
-        assertEquals("Organization", obj.getJsonObject("publisher").getString("@type"));
-        assertEquals("LibraScholar", obj.getJsonObject("publisher").getString("name"));
-        assertEquals("Organization", obj.getJsonObject("provider").getString("@type"));
-        assertEquals("LibraScholar", obj.getJsonObject("provider").getString("name"));
-        assertEquals("LibraScholar", obj.getJsonObject("includedInDataCatalog").getString("name"));
-    }
-
-    @Test
-    public void testGetJsonLdNonCC0License() throws ParseException {
-        Dataset dataset = new Dataset();
-        dataset.setProtocol("doi");
-        dataset.setAuthority("10.5072/FK2");
-        dataset.setIdentifier("LK0D1H");
-        DatasetVersion datasetVersion = new DatasetVersion();
-        datasetVersion.setDataset(dataset);
-        datasetVersion.setVersionState(DatasetVersion.VersionState.DRAFT);
-        assertEquals("", datasetVersion.getPublicationDateAsString());
-        // Only published datasets return any JSON.
-        assertEquals("", datasetVersion.getJsonLd());
-        datasetVersion.setVersionState(DatasetVersion.VersionState.RELEASED);
-        datasetVersion.setVersionNumber(1L);
-        datasetVersion.setMinorVersionNumber(0L);
-        SimpleDateFormat dateFmt = new SimpleDateFormat("yyyyMMdd");
-        Date publicationDate = dateFmt.parse("19551105");
-        datasetVersion.setReleaseTime(publicationDate);
-        dataset.setPublicationDate(new Timestamp(publicationDate.getTime()));
-        Dataverse dataverse = new Dataverse();
-        dataverse.setName("LibraScholar");
-        dataset.setOwner(dataverse);
-
-        TermsOfUseAndAccess terms = new TermsOfUseAndAccess();
-        terms.setLicense(null);
-        terms.setTermsOfUse("Call me maybe");
-        datasetVersion.setTermsOfUseAndAccess(terms);
-
-        String jsonLd = datasetVersion.getJsonLd();
-        logger.fine("jsonLd: " + JsonUtil.prettyPrint(jsonLd));
-        JsonReader jsonReader = Json.createReader(new StringReader(jsonLd));
-        JsonObject obj = jsonReader.readObject();
-        assertEquals("http://schema.org", obj.getString("@context"));
-        assertEquals("Dataset", obj.getString("@type"));
-        assertEquals("https://doi.org/10.5072/FK2/LK0D1H", obj.getString("@id"));
-        assertEquals("https://doi.org/10.5072/FK2/LK0D1H", obj.getString("identifier"));
-        assertEquals(null, obj.getString("schemaVersion", null));
-        assertTrue(obj.getString("license").contains("/api/datasets/:persistentId/versions/1.0/customlicense?persistentId=doi:10.5072/FK2/LK0D1H"));
-        assertEquals("1955-11-05", obj.getString("dateModified"));
-        assertEquals("1955-11-05", obj.getString("datePublished"));
-        assertEquals("1", obj.getString("version"));
-        // TODO: if it ever becomes easier to mock a dataset title, test it.
-        assertEquals("", obj.getString("name"));
-        // TODO: If it ever becomes easier to mock authors, test them.
-        JsonArray emptyArray = Json.createArrayBuilder().build();
-        assertEquals(emptyArray, obj.getJsonArray("creator"));
-        assertEquals(emptyArray, obj.getJsonArray("author"));
-        // TODO: If it ever becomes easier to mock subjects, test them.
-        assertEquals(emptyArray, obj.getJsonArray("keywords"));
-        assertEquals("Organization", obj.getJsonObject("publisher").getString("@type"));
-        assertEquals("LibraScholar", obj.getJsonObject("publisher").getString("name"));
-        assertEquals("Organization", obj.getJsonObject("provider").getString("@type"));
-        assertEquals("LibraScholar", obj.getJsonObject("provider").getString("name"));
-        assertEquals("LibraScholar", obj.getJsonObject("includedInDataCatalog").getString("name"));
-    }
 
 }
